Fix local deployment
The required version of `jekyll_remote_include` was incompatible
with the latest jekyll. The point is, we don't have to require a
specific version of that plugin anymore, the latest one works fine.

Also, let's not install gems into the system but rather install
them into `vendor` directory - for many obvious reasons, but mainly
to avoid conflicts between requirements for running multiple jekyll
sites on one dev system.